+++ examples/pages/Animated/index.js
@@ -1,0 +1,28 @@
+// @flow
+
+import React, { Component } from 'react';
+
+import { H1 } from '../../components';
+import ExampleWrapper from '../../ExampleWrapper';
+import Example from './Example';
+
+export default class App extends Component<*> {
+  render() {
+    return (
+      <div>
+        <H1>Animated Components</H1>
+        <p>
+          React-Select comes with Animated variants that wrap the built-in
+          components.
+        </p>
+        <p>Remove the values below to see them in action.</p>
+        <ExampleWrapper
+          label="Animation"
+          urlPath="/examples/pages/Animated/Example.js"
+        >
+          <Example />
+        </ExampleWrapper>
+      </div>
+    );
+  }
+}