+++ examples/pages/Tests.js
@@ -1,13 +1,11 @@
 // @flow
 
 import React, { Component } from 'react';
-import { withValue } from 'react-value';
 
 import Select from '../../src';
 import { H1, Hr, Note } from '../components';
 import { colourOptions, groupedOptions } from '../data';
 
-const SelectWithValue = withValue(Select);
 type State = { isDisabled: boolean, isLoading: boolean };
 
 export default class Tests extends Component<*, State> {
@@ -22,7 +20,7 @@ export default class Tests extends Component<*, State> {
         <H1>Test Page for Cypress</H1>
         <h2>Single Select</h2>
         <div id="cypress-single">
-          <SelectWithValue
+          <Select
             autoFocus
             defaultValue={colourOptions[0]}
             isDisabled={this.state.isDisabled}
@@ -48,15 +46,12 @@ export default class Tests extends Component<*, State> {
         </Note>
         <h4>Grouped</h4>
         <div id="cypress-single-grouped">
-          <SelectWithValue
-            defaultValue={colourOptions[1]}
-            options={groupedOptions}
-          />
+          <Select defaultValue={colourOptions[1]} options={groupedOptions} />
         </div>
         <Hr />
         <h2>Multi Select</h2>
         <div id="cypress-multi">
-          <SelectWithValue
+          <Select
             defaultValue={[colourOptions[2], colourOptions[3]]}
             isMulti
             options={colourOptions}