+++ src/index.js
@@ -1,6 +1,9 @@
 // @flow
-import Select from './Select';
+
+import Select from './StateManager';
 
 export default Select;
+export { default as Base } from './Select';
 export { createFilter } from './filters';
 export { components } from './components/index';
+export { mergeStyles } from './styles';