+++ src/primitives.js
@@ -20,6 +20,7 @@ const createPrimitive = (Tag: string) => ({
 export const Button = createPrimitive('button');
 export const Div = createPrimitive('div');
 export const Span = createPrimitive('span');
+export const Input = createPrimitive('input');
 
 // Assistive text to describe visual elements. Hidden for sighted users.
 export const A11yText = (props: any) => (